Restructure test folder to make it more modular
