+++ tests/integration/testcontext/context.go
@@ -1,4 +1,4 @@
-package integration
+package testcontext
 
 import (
 	"context"
@@ -11,31 +11,31 @@ import (
 // istioCrCtxKey is the key used to store the IstioCR used by a scenario in the context.Context.
 type istioCrCtxKey struct{}
 
-func getIstioCrFromContext(ctx context.Context) (*v1alpha1.Istio, bool) {
+func GetIstioCrFromContext(ctx context.Context) (*v1alpha1.Istio, bool) {
 	v, ok := ctx.Value(istioCrCtxKey{}).(*v1alpha1.Istio)
 	return v, ok
 }
 
-func setIstioCrInContext(ctx context.Context, istio *v1alpha1.Istio) context.Context {
+func SetIstioCrInContext(ctx context.Context, istio *v1alpha1.Istio) context.Context {
 	return context.WithValue(ctx, istioCrCtxKey{}, istio)
 }
 
 // testAppCtxKey is the key used to store the test app used by a scenario in the context.Context.
 type testAppCtxKey struct{}
 
-func getTestAppFromContext(ctx context.Context) (*v1.Deployment, bool) {
+func GetTestAppFromContext(ctx context.Context) (*v1.Deployment, bool) {
 	v, ok := ctx.Value(testAppCtxKey{}).(*v1.Deployment)
 	return v, ok
 }
 
-func setTestAppInContext(ctx context.Context, istio *v1.Deployment) context.Context {
+func SetTestAppInContext(ctx context.Context, istio *v1.Deployment) context.Context {
 	return context.WithValue(ctx, testAppCtxKey{}, istio)
 }
 
 // k8sClientCtxKey is the key used to store the k8sClient used during tests in the context.Context.
 type k8sClientCtxKey struct{}
 
-func getK8sClientFromContext(ctx context.Context) (client.Client, error) {
+func GetK8sClientFromContext(ctx context.Context) (client.Client, error) {
 	v, ok := ctx.Value(k8sClientCtxKey{}).(client.Client)
 	if !ok {
 		return v, errors.New("k8sClient not found in context")
@@ -43,6 +43,6 @@ func getK8sClientFromContext(ctx context.Context) (client.Client, error) {
 	return v, nil
 }
 
-func setK8sClientInContext(ctx context.Context, k8sClient client.Client) context.Context {
+func SetK8sClientInContext(ctx context.Context, k8sClient client.Client) context.Context {
 	return context.WithValue(ctx, k8sClientCtxKey{}, k8sClient)
 }
